NP-2540 Move api paths to separate file
