--- src/api/almaApi.ts
@@ -1,3 +1,0 @@
-export enum AlmaApiPaths {
-  ALMA = '/alma',
-}
