--- src/api/institutionApi.ts
@@ -1,4 +1,0 @@
-export enum InstitutionApiPaths {
-  INSTITUTIONS = '/institution/institutions',
-  DEPARTMENTS = '/institution/departments',
-}
