--- src/api/projectApi.ts
@@ -1,3 +1,0 @@
-export enum ProjectsApiPaths {
-  PROJECT = '/project',
-}
