--- src/api/searchApi.ts
@@ -1,3 +1,0 @@
-export enum SearchApiPaths {
-  REGISTRATIONS = '/search/resources',
-}
