+++ src/pages/worklist/MessagesOverview.tsx
@@ -1,7 +1,7 @@
 import { Typography } from '@material-ui/core';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { PublicationsApiPaths } from '../../api/registrationApi';
+import { PublicationsApiPaths } from '../../api/apiPaths';
 import ListSkeleton from '../../components/ListSkeleton';
 import { SupportRequest } from '../../types/publication_types/messages.types';
 import { RoleName } from '../../types/user.types';
